refactor: remove unused transaction overview components and update imports
- Deleted obsolete components related to transaction overview, including TxOverviewAmount, TxOverviewMemo, TxOverviewPanel, and TxOverviewRow, to streamline the codebase.
- Updated imports in various files to reflect the new structure and ensure consistency across the application.
