+++ clients/desktop/src/vault/keysign/join/verify/KeysignCustomMessageInfo.tsx
@@ -1,11 +1,10 @@
 import { CustomMessagePayload } from '@core/mpc/types/vultisig/keysign/v1/custom_message_payload_pb'
-import { ValueProp } from '@lib/ui/props'
-import { useTranslation } from 'react-i18next'
-
 import {
   TxOverviewChainDataRow,
   TxOverviewRow,
-} from '../../../../chain/tx/components/TxOverviewRow'
+} from '@core/ui/chain/tx/TxOverviewRow'
+import { ValueProp } from '@lib/ui/props'
+import { useTranslation } from 'react-i18next'
 
 export const KeysignCustomMessageInfo = ({
   value,
