+++ clients/desktop/src/vault/keysign/shared/KeysignSigningStep.tsx
@@ -1,4 +1,7 @@
 import { KeysignMessagePayload } from '@core/mpc/keysign/keysignPayload/KeysignMessagePayload'
+import { CurrentTxHashProvider } from '@core/ui/chain/state/currentTxHash'
+import { TxOverviewPanel } from '@core/ui/chain/tx/TxOverviewPanel'
+import { TxOverviewChainDataRow } from '@core/ui/chain/tx/TxOverviewRow'
 import { FullPageFlowErrorState } from '@core/ui/flow/FullPageFlowErrorState'
 import { useCoreNavigate } from '@core/ui/navigation/hooks/useCoreNavigate'
 import { Match } from '@lib/ui/base/Match'
@@ -16,9 +19,6 @@ import { extractErrorMsg } from '@lib/utils/error/extractErrorMsg'
 import { useEffect } from 'react'
 import { useTranslation } from 'react-i18next'
 
-import { CurrentTxHashProvider } from '../../../chain/state/currentTxHash'
-import { TxOverviewPanel } from '../../../chain/tx/components/TxOverviewPanel'
-import { TxOverviewChainDataRow } from '../../../chain/tx/components/TxOverviewRow'
 import useVersionCheck from '../../../lib/hooks/useVersionCheck'
 import { KeysignCustomMessageInfo } from '../join/verify/KeysignCustomMessageInfo'
 import { KeysignSigningState } from './KeysignSigningState'
