+++ clients/desktop/src/vault/keysign/shared/KeysignTxOverview.tsx
@@ -3,6 +3,10 @@ import { Chain } from '@core/chain/Chain'
 import { getBlockExplorerUrl } from '@core/chain/utils/getBlockExplorerUrl'
 import { fromCommCoin } from '@core/mpc/types/utils/commCoin'
 import { KeysignPayload } from '@core/mpc/types/vultisig/keysign/v1/keysign_message_pb'
+import { useCurrentTxHash } from '@core/ui/chain/state/currentTxHash'
+import { TxOverviewAmount } from '@core/ui/chain/tx/TxOverviewAmount'
+import { TxOverviewMemo } from '@core/ui/chain/tx/TxOverviewMemo'
+import { TxOverviewRow } from '@core/ui/chain/tx/TxOverviewRow'
 import { useOpenUrl } from '@core/ui/state/openUrl'
 import { IconButton } from '@lib/ui/buttons/IconButton'
 import { CopyIcon } from '@lib/ui/icons/CopyIcon'
@@ -18,11 +22,7 @@ import { matchDiscriminatedUnion } from '@lib/utils/matchDiscriminatedUnion'
 import { useMemo } from 'react'
 import { useTranslation } from 'react-i18next'
 
-import { useCurrentTxHash } from '../../../chain/state/currentTxHash'
 import { nativeSwapChains } from '../../../chain/swap/native/NativeSwapChain'
-import { TxOverviewAmount } from '../../../chain/tx/components/TxOverviewAmount'
-import { TxOverviewMemo } from '../../../chain/tx/components/TxOverviewMemo'
-import { TxOverviewRow } from '../../../chain/tx/components/TxOverviewRow'
 import { formatFee } from '../../../chain/tx/fee/utils/formatFee'
 import { useCopyTxHash } from '../../../chain/ui/hooks/useCopyTxHash'
 import { useCoinPriceQuery } from '../../../coin/query/useCoinPriceQuery'
