+++ clients/desktop/src/vault/keysign/shared/KeysignTxPrimaryInfo.tsx
@@ -2,6 +2,13 @@ import { fromChainAmount } from '@core/chain/amount/fromChainAmount'
 import { Chain } from '@core/chain/Chain'
 import { fromCommCoin } from '@core/mpc/types/utils/commCoin'
 import { KeysignPayload } from '@core/mpc/types/vultisig/keysign/v1/keysign_message_pb'
+import { TxOverviewAmount } from '@core/ui/chain/tx/TxOverviewAmount'
+import { TxOverviewMemo } from '@core/ui/chain/tx/TxOverviewMemo'
+import {
+  TxOverviewChainDataRow,
+  TxOverviewPrimaryRowTitle,
+  TxOverviewRow,
+} from '@core/ui/chain/tx/TxOverviewRow'
 import { ValueProp } from '@lib/ui/props'
 import { MatchQuery } from '@lib/ui/query/components/MatchQuery'
 import { shouldBePresent } from '@lib/utils/assert/shouldBePresent'
@@ -10,13 +17,6 @@ import { assertField } from '@lib/utils/record/assertField'
 import { useMemo } from 'react'
 import { useTranslation } from 'react-i18next'
 
-import { TxOverviewAmount } from '../../../chain/tx/components/TxOverviewAmount'
-import { TxOverviewMemo } from '../../../chain/tx/components/TxOverviewMemo'
-import {
-  TxOverviewChainDataRow,
-  TxOverviewPrimaryRowTitle,
-  TxOverviewRow,
-} from '../../../chain/tx/components/TxOverviewRow'
 import { formatFee } from '../../../chain/tx/fee/utils/formatFee'
 import { useCoinPriceQuery } from '../../../coin/query/useCoinPriceQuery'
 import { useFiatCurrency } from '../../../preferences/state/fiatCurrency'
