+++ clients/desktop/src/vault/keysign/shared/SwapKeysignTxOverview.tsx
@@ -4,6 +4,7 @@ import { getBlockExplorerUrl } from '@core/chain/utils/getBlockExplorerUrl'
 import { fromCommCoin } from '@core/mpc/types/utils/commCoin'
 import { OneInchSwapPayload } from '@core/mpc/types/vultisig/keysign/v1/1inch_swap_payload_pb'
 import { KeysignPayload } from '@core/mpc/types/vultisig/keysign/v1/keysign_message_pb'
+import { useCurrentTxHash } from '@core/ui/chain/state/currentTxHash'
 import { useCoreNavigate } from '@core/ui/navigation/hooks/useCoreNavigate'
 import { useOpenUrl } from '@core/ui/state/openUrl'
 import { useCurrentVault } from '@core/ui/vault/state/currentVault'
@@ -15,6 +16,7 @@ import { sameDimensions } from '@lib/ui/css/sameDimensions'
 import { ChevronRightIcon } from '@lib/ui/icons/ChevronRightIcon'
 import { SquareArrowTopIcon } from '@lib/ui/icons/SquareArrowTopIcon'
 import { AnimatedVisibility } from '@lib/ui/layout/AnimatedVisibility'
+import { SeparatedByLine } from '@lib/ui/layout/SeparatedByLine'
 import { HStack, VStack } from '@lib/ui/layout/Stack'
 import { ValueProp } from '@lib/ui/props'
 import { GradientText, Text } from '@lib/ui/text'
@@ -26,9 +28,7 @@ import { useMemo } from 'react'
 import { useTranslation } from 'react-i18next'
 import styled, { css } from 'styled-components'
 
-import { useCurrentTxHash } from '../../../chain/state/currentTxHash'
 import { formatFee } from '../../../chain/tx/fee/utils/formatFee'
-import { SeparatedByLine } from '../../../lib/ui/layout/SeparatedByLine'
 import { SwapCoinItem } from './SwapCoinItem'
 
 export const SwapKeysignTxOverview = ({ value }: ValueProp<KeysignPayload>) => {
