+++ clients/desktop/src/vault/swap/keysign/KeysignSwapTxInfo.tsx
@@ -1,6 +1,10 @@
 import { fromChainAmount } from '@core/chain/amount/fromChainAmount'
 import { Chain } from '@core/chain/Chain'
 import { KeysignPayload } from '@core/mpc/types/vultisig/keysign/v1/keysign_message_pb'
+import {
+  TxOverviewChainDataRow,
+  TxOverviewRow,
+} from '@core/ui/chain/tx/TxOverviewRow'
 import { ValueProp } from '@lib/ui/props'
 import { withoutUndefined } from '@lib/utils/array/withoutUndefined'
 import { formatTokenAmount } from '@lib/utils/formatTokenAmount'
@@ -14,10 +18,6 @@ import {
   toKeysignSwapPayload,
 } from '../../../chain/keysign/KeysignSwapPayload'
 import { generalSwapProviderName } from '../../../chain/swap/general/GeneralSwapProvider'
-import {
-  TxOverviewChainDataRow,
-  TxOverviewRow,
-} from '../../../chain/tx/components/TxOverviewRow'
 
 export const KeysignSwapTxInfo = ({ value }: ValueProp<KeysignPayload>) => {
   const { erc20ApprovePayload, toAmount: fromAmount } = value
