+++ core/ui/chain/state/currentTxHash.ts
