+++ core/ui/chain/tx/TxOverviewPanel.tsx
@@ -1,7 +1,6 @@
+import { SeparatedByLine } from '@lib/ui/layout/SeparatedByLine'
 import { Panel } from '@lib/ui/panel/Panel'
 import { ChildrenProp } from '@lib/ui/props'
-
-import { SeparatedByLine } from '../../../lib/ui/layout/SeparatedByLine'
 
 export const TxOverviewPanel: React.FC<ChildrenProp> = ({ children }) => (
   <Panel>
